refactor: Move DevTokens DTOs to plugin namespace
