--- Infrastructure/Restful/DeveloperTokens/DTO/CreateTokenRequest.cs
@@ -1,8 +1,0 @@
-﻿namespace Infrastructure.Restful.DeveloperTokens.DTO;
-
-public record CreateTokenRequest(
-    string Name,
-    string Description,
-    IEnumerable<string> Scopes,
-    int? LifetimeDays
-);
